Added custom size factor support and tests
Upscalers :
 - bilinear : supports custom size factor
 - nearest : now supports custom size factor
 - waifu2x_ncnn_vulkan : supports both custom size and denoising, this one is just a wrapper for real waifu2x

Made the code more adaptable, it is now possible to make a wrapper for external upscaler (example on waifu2x_ncnn_vulkan) and adapted the not yet working video upscaler to these changes.

I also added a simpler way to test upscalers, go and check in tests.py for details.