Implemented algorithm in C++
For performance reasons, everything is not implemented in C++ in addition of Python, it can be up to 100 times faster.
Colour spaces are not guaranteed to work outside of RGB 8bits, so be careful about this, more investigations needed there.

Added a build script for C++ upscalers, run  it from project's root likewise tests script.

Added a deps list, updated tests script to latest algorithms and added wrappers for C++ programs.

Also fixed some bugs and
ImageMagick got yeeted.